chore: move dot to library
