+++ src/lib/components/svgs/devx.svelte
@@ -1,5 +1,5 @@
 <script>
-	import Dot from '../dot.svelte';
+	import Dot from '../library/dot.svelte';
 </script>
 
 <div class="flex items-center">
